cpufreq: Resolve CPUFREQ_NOTIFY issue
When the cpufreq_stat receives the frequency policy notification,
it checks if the frequency table and stats table are present for the cpu.
If both are present, it returns with -EBUSY error to the notification.
Due to this, none of the other registered modules receive this
notification.

If the stat table and sysfs nodes are already created, return 0 instead
of returning EBUSY error.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
(cherry picked from commit 5893904eac05642e5c416f6ce529a39a0be2417b)
(cherry picked from commit 587c3340009e277956466930c9534244c4dd9af4)